[REFAC: #19] - Reorganizar os testes.
close #19
